整理了项目文件

+++ linked_list_test.c
@@ -1,3 +1,5 @@
+#include <stdio.h>
+
 #include "linked_list.h"
 
 int main()
@@ -11,5 +13,5 @@ int main()
     // for(int i=0; i<10; i++) printf("%lu\n", list_index(head, i));
 
     printf("index:%lu\n", list_in(head, (void *)10));
-    printf("list len:%lu\n", list_len(head));
+    printf("list len:%lu\n", (unsigned long)list_len(head));
 }
